# backend/API/swagger.yaml
swagger: '3.0.0'
info:
  title: SmartSupply API
  description: REST API for managing product tracking using blockchain technology
  version: '1.0.0'
basePath: /api
schemes:
  - http
  - https
...
paths:
...
  /escrow/create-lot-listing:
    post:
      tags:
        - Escrow
      summary: Create a new lot listing
      description: Creates a new listing for a lot in the escrow system
      security:
        - cookieAuth: []
      parameters:
        - in: body
          name: body
          description: Batch ID, Lot ID, and Price Per Unit for the lot listing
          required: true
          schema:
            type: object
            properties:
              batchId:
                type: integer
              lotId:
                type: integer
              pricePerUnit:
                type: number
      responses:
        200:
          description: Successful operation
          schema:
            type: object
            properties:
              message:
                type: string
        400:
          description: Bad request
          schema:
            type: object
            properties:
              error:
                type: string
      produces:
        - application/json
  /escrow/create-batch-listing:
    post:
      tags:
        - Escrow
      summary: Create a new batch listing
      description: Creates a new listing for a batch in the escrow system
      security:
        - cookieAuth: []
      parameters:
        - in: body
          name: body
          description: Batch ID and Price Per Unit for the batch listing
          required: true
          schema:
            type: object
            properties:
              batchId:
                type: integer
              pricePerUnit:
                type: number
      responses:
        200:
          description: Successful operation
          schema:
            type: object
            properties:
              message:
                type: string
        400:
          description: Bad request
          schema:
            type: object
            properties:
              error:
                type: string
      produces:
        - application/json
  /escrow/purchase-lot:
    post:
      tags:
        - Escrow
      summary: Purchase a lot
      description: Purchases a lot from the escrow system
      security:
        - cookieAuth: []
      parameters:
        - in: body
          name: body
          description: Listing ID and amount for the lot purchase
          required: true
          schema:
            type: object
            properties:
              listingId:
                type: integer
              amount:
                type: number
      responses:
        200:
          description: Successful operation
          schema:
            type: object
            properties:
              message:
                type: string
        400:
          description: Bad request
          schema:
            type: object
            properties:
              error:
                type: string
      produces:
        - application/json
  /escrow/purchase-batch:
    post:
      tags:
        - Escrow
      summary: Purchase a batch
      description: Purchases a batch from the escrow system
      security:
        - cookieAuth: []
      parameters:
        - in: body
          name: body
          description: Listing ID and amount for the batch purchase
          required: true
          schema:
            type: object
            properties:
              listingId:
                type: integer
              amount:
                type: number
      responses:
        200:
          description: Successful operation
          schema:
            type: object
            properties:
              message:
                type: string
        400:
          description: Bad request
          schema:
            type: object
            properties:
              error:
                type: string
      produces:
        - application/json
securityDefinitions:
  cookieAuth:
    type: apiKey
    in: header
    name: connect.sid